+++ src/lib/DataTableHeader/DataTableRow.tsx
@@ -32,10 +32,14 @@ export function DataTableRow<T, TRouteNames>({ keyField, record, columns, action
               />
             )}
             {actions.view && (
-              <actions.view.link legacyBehavior route={actions.view.route} params={actions.view.getParams({ keyValue, cell: record })}>
-                <a>
+              <actions.view.link route={actions.view.route} params={actions.view.getParams({ keyValue, cell: record })}>
+                {actions.view.omitHyperLink ? (
                   <FontAwesomeIcon icon={faEye} style={{ marginRight: "5px" }} />
-                </a>
+                ) : (
+                  <a>
+                    <FontAwesomeIcon icon={faEye} style={{ marginRight: "5px" }} />
+                  </a>
+                )}
               </actions.view.link>
             )}
             {actions.delete && (